sdhc: boot powerpc code with sdhc quiesced
This patch changes 'mini' to quiesce the Secure Digital Host Controller
Interface before booting PowerPC code. This is needed to avoid having to
implement a 'mini' IPC layer on the PowerPC side just to tell 'mini' that
the SDHC hardware is going to be managed from the Broadway processor.

This patch also ensures that 'mini' regains control of the SDHC hardware
when a call to IPC_SDHC_DISCOVER or the first SDMMC call is performed.
This way, no changes are needed on Ceiling Cat or HBC.

Signed-off-by: Albert Herranz <albert_herranz@yahoo.es>